이메일 보내기 기능 추가

+++ src/main/java/com/spring/model/ExampleUserRepository.java
@@ -2,6 +2,6 @@ package com.spring.model;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ExampleUserRepository extends JpaRepository<ExampleUser, Integer> {
+public interface ExampleUserRepository extends JpaRepository<ExampleUser, Long> {
     ExampleUser findByUserLoginId(String loginId);
 }
